Updates sprites for all game objects, added beat markers on main axis for better readability
Enemys and bullets now have a direction on their sprite to each with reading.
Enemies also have a center and center axis marking for better timing.
The flower has been redesigned to show not only the perfect timing zone, but also the tracks which notes can approach some. Might still redesign this later.
Dynamic beat markings are also added to the scene based on the speed of notes and temp of songs. This should help improve readability and timing as well.